chore: refactor mqttPacket
